requirements.txt
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
filelock==3.13.3
Flask==3.0.2
Flask-Cors==4.0.0
fsspec==2024.3.1
gunicorn==21.2.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
packaging==24.0
pillow==10.3.0
requests==2.31.0
sympy==1.12
torch==2.2.2
torchvision==0.17.2
typing_extensions==4.10.0
urllib3==2.2.1
Werkzeug==3.0.2